+++ src/PairwiseListMatrices.jl
@@ -8,7 +8,7 @@ module PairwiseListMatrices
   import Base: size, getindex, setindex!, length, eltype,
   similar, copy, abs, full, zeros, ones, convert,
   transpose, transpose!, ctranspose, ctranspose!, diag, svd,
-  triu!, triu, writedlm, writecsv,
+  triu!, triu, writedlm, writecsv, join,
   -, +, .*, .-, .+, .-, ./, *, /,
   # faster
   mean, sum, varm, var, std, sqrt